requirements.txt
click>=8.1.7
numpy>=2.0.0
pyyaml>=6.0.2
orjson>=3.10.7
jsonschema>=4.22.0
openai>=1.43.0
pydub>=0.25.1